Insert file in DB only once
This commit has many changes being the main one the fact that we insert
a file in the DB only once. Now we calculate the sha1 of the file and
check if it is compatible with the one in the DB. If it is, there is
nothing else to do. If it is not, we remove all the references to this
file, insert a fresh file row and give the signal to parse the file.

In the parsing, now we make sure to only insert the parsing file
(including headers) only once. Also, I changed the parsing argument to
get nodes of pre-processor statements.

In the DB, I added a a symbol_decls table and made func_defs to point to
the new table (given the definition, we want to find the declaration.)
The new table will contain the declaration of all the symbols including
functions and variables. Also, the structure Function was renamed
Symbol.

I also added a text on how we can augment the database for more accurate
symbol declaration searches (with a symbols table.)