Added table for n-to-n relationship between func defs and decls
Function declarations and definitions have a n-to-n relationship. One
definition can have multiple declarations (it is valid to declare one function
several times) and one declaration can have multiple definitions (e.g. a header
file defining an interface). This path add to the database the capability of
supporting this relationships. Also, the method to insert a function
declaration with its definition is implemented.