Removed func_defs table in favor of symbol_decls
I realized that the definition of a function is nothing but a
specialized declaration. I removed the definition table and added a
'def' field in the declaration table. Now the joining table
'func_decs_defs' points twice to the symbol_decls and the definition
declarations will point to itself. This will help in functions with no
other declaration but its own definition.

I also removed the name as part of the primary key as I realized there
could not be another symbol in the same position. This saves code (as
the foreign keys are shorter) and space in the DB.

I added the USR as part of the symbol info. This will help when a
definition of a declaration is not found (resolved at linking time) in a
query.