Export all memeber of Symbols to fix JSONRPC mechanism
For the marshalling to work properly, the members of the marshalled
structure should be exported. This change exports all these members. To
test, I added a remote method to return the declaration of a function
use.

Added sample call in the test client.go. Next we should include a
python client as this is very likely to be used in a editor plugin.

Also added a comment of the issues with the automatic updating and the
dependency between header files and its implementations.