Handling case where the symbol is not found
Whenever the symbols is not found, the find declaration function returns
nil. I was assigning directly to the returned pointer and this was
causing a crash. Now the appropriate error is returned.

BONUS: (1) Add end-of-string to the matchin regular expression of C
files and (2) Remove socket file before trying to open the new one.