Adding in-memory DB for faster execution
Instead of having the DB in disk, this patch changes the DB to be in
memory instead. This speeds-up the indexing massively (clearly the DB
was a huge bottleneck before.) Additionally, to not loose all the info
in case of crash, the information is flushed to disk every 30 seconds.
Moreover, the DB is also flushed when the daemon shuts down cleanly (so
we don't have to re-index the whole code again.)

BONUS: (1) Moving WaitGroup.Add calls inside the threads instead and (2)
Adding *.swp to gitignore.