Wrapping single-file parsing DB insertions in a single transaction
In the previous code, I was using the golang sql interface the wrong
way. This was causing issues with the memory and with the performance.
This patch introduces the use of a transaction for all the insertions in
a file parsing.

Unfortunately, for now, this only supports a single thread modifying the
DB at the time. Otherwise, we get database/table lock errors. This
should be eventually fixed.