Making sure file and compile options match abs/relative path in input
In order for headers file path to match input directories, we need to
make sure that the include directories match either absolute or relative
path of input directory. This helps to keep consistency in the DB with a
single name for the same header file.

BONUS: Now that we have compilation options, it does not make much sense
to explore header files by itself. They won't have the compilation
options to know exactly what is getting used. In fact, if the header
file is not included by any C file, it won't be compiled. This should be
easy to change if we find other benefit to parse header files alone.