Using mxk/go-sqlite driver instead
The new sqlite driver gives you much better control over the sqlite
connections and the internals of the database. I divided the database in
writers and readers now. This will allow to consume the writer memory
only when necessary.

This is making the code runs faster than it used to. Currently OpenBSD
kernel is indexed in 1m52s while linux takes ~14m both with one single
thread. This should improve with multiple threads.