When a header file changes, removing all its dependents and reparse
In this commit I move all the code to handle files to files.go. This
includes scanning the files to index, watching the files for change, and
checking if files are up-to-date before parsing and when inserting a new
dependency.

For the watchers I removed the broken watching of directories. I will
assume for now this is not very common and leave the feature for later.