Add method to find symbol definition
This patch adds a method to find the symbol definition. This is for now
working with functions only as we are not indexing structures yet. If
navc knows the definition for sure, it will return only this definition.
The only requirement for this to happen is that the declaration and the
definition are compiled together in the same translation unit. This
should be very common (otherwise, the code is not well structured).

If this is not possible, it will return all the definitions with the
same USR. This should be very reliable too, returning in the worst case
a very narrow list.