Change concurrency model to fix race conditions
Before, the code was full of race conditions. The watcher could fire on
a file and do changes while a parser was parsing the same file and doing
changes. In this patch I change the concurrency model of the
application. We have a goroutine that handle all files changes while
many goroutines parse concurrently several files. While parsing, we
simply create a parsed index object that the file handler goroutine will
check and insert to the DB if everything looks good. When starting, a
goroutine explores the indexed directory looking for new and erased
files and feed the file handler goroutine to handle this cases
correctly. With this change, and 12 cores, the indexer is able to index
all the Linux kernel in under 3 minutes. The index size is about 1.4GB.