Have temp tudb file inside our dbDir
Having the temp tudb in /tmp/ was making the Move operation to fail
(because it was not in the same file system). This patch creates a
temporary directory in our index directory for this purpose. As a side
effect, dbDirPath is now a global variable. This is OK as we will have
one DB per daemon.