Add support to non-existing headers
If a header does not exist when a file is parsed, we create a
placeholder in the DB in case the file shows up. If any header is
created, we look for that filename in the DB for potential includers. If
we find any, we reparse them in case the file finally showed up.

Also, this patch fixes a bug where we were not handling the create and
rename events.